feat: refactor longitude handling and improve flight tracking logic
